Fix HVAC temperature signals' data types
VSS 5.0 aligned all temperature signal data types to float, so our
use of the HVAC temperature signals needs to be updated to match.

Bug-AGL: SPEC-5386

+++ lib/data/data_providers/vehicle_notifier.dart
@@ -165,17 +165,17 @@ class VehicleNotifier extends Notifier<Vehicle> {
     try {
       switch (side) {
         case Side.left:
-          valClient.setInt32(
+          valClient.setFloat(
             VSSPath.vehicleDriverTemperature,
-            value,
+            value.toDouble(),
             true,
           );
           state = state.copyWith(driverTemperature: value);
           break;
         case Side.right:
-          valClient.setInt32(
+          valClient.setFloat(
             VSSPath.vehiclePassengerTemperature,
-            value,
+            value.toDouble(),
             true,
           );
           state = state.copyWith(passengerTemperature: value);
